DC-417 Deid_02
This allows deid generalizations to work for the id based
configuration files.  The config files are updated to
defined requirements.
adds 'like' and 'not like' row suppression abilities
NOT IN cannot be used with nullable columns.  Should instead use
NOT EXISTS.
The config files were adjusted and the press and rules files
were adjusted to allow using exists/not exists in the config
files to search/update on nullable columns.